+++ slash/src/main/java/project/slash/statistics/repository/StatisticsRepositoryCustom.java
@@ -1,15 +1,11 @@
 package project.slash.statistics.repository;
 
+import java.time.LocalDate;
 import java.util.List;
 
 import project.slash.statistics.dto.response.MonthlyDataDto;
-import project.slash.statistics.dto.response.MonthlyServiceStatisticsDto;
 import project.slash.statistics.dto.response.MonthlyStatisticsDto;
 import project.slash.statistics.dto.response.StatisticsDto;
-import project.slash.statistics.dto.request.RequestStatisticsDto;
-import project.slash.statistics.dto.MonthlyDataDto;
-import project.slash.statistics.dto.MonthlyServiceStatisticsDto;
-import project.slash.statistics.dto.StatisticsDto;
 import project.slash.statistics.dto.response.ResponseServiceTaskDto;
 
 public interface StatisticsRepositoryCustom {
